requirements.txt
Pillow==9.4.0
Pillow==9.5.0
Requests==2.28.2
streamlit==1.20.0
streamlit_authenticator==0.1.5
streamlit_lottie==0.0.3
~treamlit==1.20.0
~treamlit==1.3.0
